watchdog indexer
