+++ indexer/cmd/orchestrator.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"net/http"
-	"time"
 
 	"github.com/coreos/go-systemd/v22/daemon"
 	config "github.com/mezonai/mmn-tx-explorer/indexer/configs"
@@ -59,24 +58,4 @@ func RunOrchestrator(cmd *cobra.Command, args []string) {
 	}
 
 	orchestratorService.Start()
-}
-
-func startSystemdWatchdog() {
-	interval, err := daemon.SdWatchdogEnabled(false)
-	if err != nil || interval == 0 {
-		return
-	}
-
-	log.Info().
-		Dur("interval", interval).
-		Msg("Systemd watchdog enabled")
-
-	go func() {
-		ticker := time.NewTicker(interval / 2)
-		defer ticker.Stop()
-
-		for range ticker.C {
-			daemon.SdNotify(false, daemon.SdNotifyWatchdog)
-		}
-	}()
 }
